Move HDF5 support to a package extension
HDF5 is now a weakdep; all h5save/h5load/save/load methods that take HDF5
types live in ext/HDF5Ext.jl and are loaded automatically when the user
loads HDF5.

Users no longer need HDF5 present to load AbstractNeuralNetworks; it is
only required when saving or loading NeuralNetworkParameters.

- HDF5 moved from [deps] to [weakdeps] with a corresponding [extensions] entry
- using HDF5 / using HDF5: H5DataStore removed from the main module
- HDF5-specific functions removed from parameters.jl; replaced by stubs
  (function h5save end etc.) so downstream packages can extend them
- All HDF5 implementations moved to ext/HDF5Ext.jl
- HDF5 added to [extras]/[targets] so the test suite can still use it
