+++ Project.toml
@@ -4,11 +4,14 @@ authors = ["Michael Kraus"]
 version = "0.6.1"
 
 [weakdeps]
+HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
+
+[extensions]
+HDF5Ext = "HDF5"
 
 [deps]
 GPUArraysCore = "46192b85-c4d5-4398-a991-12ede77f4527"
-HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
 KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
@@ -25,10 +28,11 @@ ZygoteRules = "0.2.7"
 julia = "1.6"
 
 [extras]
+HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Random", "LinearAlgebra", "SafeTestsets", "Test", "Zygote"]
+test = ["HDF5", "Random", "LinearAlgebra", "SafeTestsets", "Test", "Zygote"]
